ti-sgx-ddk-um: avoid unnecessary dependency on -dev package
Source: meta-ti
MR: 129373
Type: Integration
Disposition: Merged from meta-ti
ChangeID: 7e4cdc3c15a9ab0dd9355a811cd9d9cb911f2765
Description:

ti-sgx-ddk-um's do_install creates a symlink libGLESv2.so.1 to
libGLESv2.so. As this symlink will be installed to ti-sgx-ddk-um, but
its target libGLESv2.so is in ti-sgx-ddk-um-dev, the automated
dependency analysis will add ti-sgx-ddk-um-dev to ti-sgx-ddk-um's
RDEPENDS. This will pull in a large number of other -dev packages
through RDEPENDS and RRECOMMENDS in turn.

Avoid this by having libGLESv2.so.1 point to libGLESv2.so.2 instead of
libGLESv2.so, so the symlink's target is in ti-sgx-ddk-um as well.

Signed-off-by: Matthias Schiffer <matthias.schiffer@ew.tq-group.com>
Reviewed-by: Denys Dmytriyenko <denys@konsulko.com>
Signed-off-by: Ryan Eatmon <reatmon@ti.com>
Signed-off-by: Denys Dmytriyenko <denys@konsulko.com>
Signed-off-by: Ryan Eatmon <reatmon@ti.com>
Signed-off-by: Jeremy A. Puhlman <jpuhlman@mvista.com>