optee: make bbappends version-specific
Source: meta-ti
MR: 129373
Type: Integration
Disposition: Merged from meta-ti
ChangeID: d59d441e1add7f71f71645518e139d855492f91e
Description:

Kirkstone branch of meta-arm comes with 2 versions of optee-os and
optee-client - 3.14 and 3.16. They have different dependencies and
slightly different build flows.

In our bbappends we bump SRCREV to latest version for newer platform
support and also update PV to indicate the correct version. While
bbappends are expected to be applied against newer 3.16 version of the
recipes, they use wildcard versioning and are not version-specific.

That results in bbappends applied to both versions of the recipes and
PV set to be the same, while the content is different. At this point
bitbake can pick up any version of the base recipe and if it gets
3.14 to apply the bbappend, the build fails due to that version being
too old and have different dependencies and build flows.

Making bbappends version-specific avoids the problem. Alternatively,
a whole recipe of the required newer version could be added, instead
of using bbappends.

Also, this is specific to Kirkstone, where meta-arm provides two
versions of optee - 3.14 and 3.16. In Lagdale and master only a
single 3.18 version of optee is provided.

Another hacky approach would be to not bump PV and let it be set by
the base recipe. That would avoid the build issue and keep bbappends
version-agnostic across branches, but result in incorrect versioning
of the binary packages, which might be confusing.

Signed-off-by: Denys Dmytriyenko <denys@konsulko.com>
Signed-off-by: Ryan Eatmon <reatmon@ti.com>
Signed-off-by: Jeremy A. Puhlman <jpuhlman@mvista.com>