linux-ti-*: Add filtering for which dtb/dtbo to include
Source: meta-ti
MR: 129373
Type: Integration
Disposition: Merged from meta-ti
ChangeID: aacaba30db61d9199a23513ef6838922d73dcad0
Description:

We often carry more dtbs/dtbos in our kernel than we have upstreamed.
The inclusion of all of the dtb/dtbo in the KERNEL_DEVICETREE has become
problematic as we start testing the linux-next and 6.1 as part of our
LTS migration.  To address this issue we are adding in two step method
for managing the KERNEL_DEVICETREE.

Going forward we will only accept dtb/dtbo in KERNEL_DEVICETREE if it is
available in upstream.  This way we ensure that the variable is more
accurate for whichever kernel you might be looking at.  We have also
added a new variable KERNEL_DEVICETREE_PREFIX which our kernel recipes
will use to auto set KERNEL_DEVICETREE based on what files are in the
kernel and not a fixed list in the conf files.

Signed-off-by: Ryan Eatmon <reatmon@ti.com>
Signed-off-by: Andrew Davis <afd@ti.com>
Signed-off-by: Ryan Eatmon <reatmon@ti.com>
Signed-off-by: Jeremy A. Puhlman <jpuhlman@mvista.com>