u-boot-ti-staging: Add 2023.04 in prep for LTS migration
Source: meta-ti
MR: 129373
Type: Integration
Disposition: Merged from meta-ti
ChangeID: 57da1808a3cd28b548b58621c0349220d9b13e6b
Description:

We are preparing to migrate our u-boot version to 2023.04.  As we are still
working on the migration, mark the default preference low for this
version.  We will enable building with this u-boot in a meta-arago brand
in the near future.  Once the migration is complete, we will remove the
2021.01 u-boot and make this the default.

Signed-off-by: Ryan Eatmon <reatmon@ti.com>
Signed-off-by: Praneeth Bajjuri <praneeth@ti.com>
Signed-off-by: Jeremy A. Puhlman <jpuhlman@mvista.com>