conf: Remove vmlinux.gz from KERNEL_IMAGETYPES
Source: meta-ti
MR: 129373
Type: Integration
Disposition: Merged from meta-ti
ChangeID: e64e97b1c128352818636c5a437b0abafeec83fc
Description:

With a recent patch to the TI kernel that includes debug symbols,
the vmlinux.gz file grew in size that impacted the testing.
This file is not needed to be in the final image so remove it
from the list.

Signed-off-by: Ryan Eatmon <reatmon@ti.com>
Signed-off-by: Denys Dmytriyenko <denys@konsulko.com>
Signed-off-by: Ryan Eatmon <reatmon@ti.com>
Signed-off-by: Jeremy A. Puhlman <jpuhlman@mvista.com>