u-boot-ti-staging: Fix u-boot configs to match oe-core fitImage
Source: meta-ti
MR: 129373
Type: Integration
Disposition: Merged from meta-ti
ChangeID: d484ef5b8196206cc60e5c099da659dca9022dd4
Description:

After moving to the oe-core kernel-fitimage.bbclass we found that the
dtb names we were searching for did not match.  This was due to the
KERNEL_DEVICETREE entries including the vendor subdir in the section
names of the fitImage file.  This patch updates support in 2021 so that
we can temporarily get to a stable boot point before starting our LTS
migration and changing a lot of things in the recipes.

Signed-off-by: Ryan Eatmon <reatmon@ti.com>
Signed-off-by: Jeremy A. Puhlman <jpuhlman@mvista.com>