kernel-fitimage: Add signing of fitImage entries for 5.10 kernel
Source: meta-ti
MR: 129373
Type: Integration
Disposition: Merged from meta-ti
ChangeID: 183e6abac6611fea0f52d7ac7303131536516a80
Description:

While we are mirgrating to kernel v6.1 and u-boot v2023.04, we want to
keep kernel v5.10 and u-boot v2021.01 "working" for anyone looking at
kirkstone.  One of the items we are looking at changing is the signing
of entries in the fitImage.

To try and acheive a limited parity with dunfell while we work on the
migration, this commit creates a class that implements the logic that was in
dunfell and applies that class to the 5.10 kernel.

This is a temporary patch that will be remvoed when we remove the 5.10
recipe down the road.  This logic will not apply to the 6.1 kernel.

Signed-off-by: Ryan Eatmon <reatmon@ti.com>
Signed-off-by: Jeremy A. Puhlman <jpuhlman@mvista.com>