meta-ti-bsp: sdimage-2part-efi: rename to load variables
Source: meta-ti
MR: 129373
Type: Integration
Disposition: Merged from meta-ti
ChangeID: 9002a4e3607503fe7e4ba8aab4b428159858e8f4
Description:

Rename the .wks to .wks.in so the EFI_PROVIDER variable is expanded
properly. We only ever set a soft recommendation for GRUB. If anyone
attempted to change that previously it was ignored (for more than just
the reason here, but that's out of my control).

Add a timeout for the bootloader menu. Currently setting it to 3
seconds.

Remove the read only flag on the rootfs. Not sure why that was there.

Signed-off-by: Randolph Sapp <rs@ti.com>
Signed-off-by: Ryan Eatmon <reatmon@ti.com>
Signed-off-by: Jeremy A. Puhlman <jpuhlman@mvista.com>