kernel_devicetree_vendor_dtb: Backport kernel-devicetree.bbclass
Source: meta-ti
MR: 129373
Type: Integration
Disposition: Merged from meta-ti
ChangeID: bdc8e7c9b58be79cd94906d1a2b6571e61d35561
Description:

There is a series of patches [1] [2] [3] that were submitted to oe-core
master that have gone through some iterations.  It has partially stalled
in the upstreaming process but we need the effective changes to continue
our LTS plans for u-boot 2023.

This backports the effective changes of that patch against the kirkstone
version of the two classes, but renames them to make it obvious that we
are overriding them.  This should be a temporary change as we expect the
changes to be backported to oe-core kirkstone once the stall is gone.
This patch will be removed once it makes it into oe-core kirkstone.

[1] https://lists.openembedded.org/g/openembedded-core/message/180753
[2] https://lists.openembedded.org/g/openembedded-core/message/181190
[2] https://lists.openembedded.org/g/openembedded-core/message/181191

Signed-off-by: Randolph Sapp <rs@ti.com>
Signed-off-by: Ryan Eatmon <reatmon@ti.com>
Signed-off-by: Jeremy A. Puhlman <jpuhlman@mvista.com>