mesa-pvr: add dynamic-layer append for llvmpipe
Source: meta-ti
MR: 129373
Type: Integration
Disposition: Merged from meta-ti
ChangeID: 0271f0a9bc355abcea41ea06be9fddc598a0d7d1
Description:

Meta-clang provides a bbappend for mesa to use clang to accelerate it's
rasterization with runtime code generation through the use of llvmpipe.
With the addition of mesa-pvr we no longer get this for free with the
blanket mesa bbappend on the devices that explicitly request mesa-pvr so
lets add a dynamic layer to append the same args meta-clang does.

Though the number of devices that would select pvr-mesa and still want
to use software rendering is small, it's not zero due to debug and
testing purposes.

Signed-off-by: Randolph Sapp <rs@ti.com>
Signed-off-by: Ryan Eatmon <reatmon@ti.com>
Signed-off-by: Jeremy A. Puhlman <jpuhlman@mvista.com>