recipes-security: optee: Remove software PRNG for am62pxx
With the latest TIFS fix [1], AM62P is able to leverage the hardware-TRNG
instead of having to fallback into the software-PRNG.

Background:
The default priv-id for A53 cores in SYSFW for AM62P was chosen to be 4
while the other K3 SOCs were configured at 1. Any firewall TISCI API
call that used 1 would not be able to run on AM62P. The TIFS fix changes
the AM62P value from 4 to 1.

[1] -
https://git.ti.com/cgit/processor-firmware/ti-linux-firmware/commit/?h=ti-linux-firmware&id=98de0a4290ec200cc224d4a5a6f42d2126230e2f

Signed-off-by: Aashvij Shenai <a-shenai@ti.com>
Acked-by: Chirag Shilwant <c-shilwant@ti.com>